+++ lib/generators/paper_trail/install_generator.rb
@@ -43,10 +43,13 @@ module PaperTrail
     end
 
     def create_initializer
-      create_file(
-        "config/initializers/paper_trail.rb",
-        "PaperTrail.config.track_associations = #{!!options.with_associations?}\n"
-      )
+      if options.with_associations?
+        create_file(
+          "config/initializers/paper_trail.rb",
+          "PaperTrail.config.track_associations = true\n",
+          "PaperTrail.config.association_reify_error_behaviour = :error"
+        )
+      end
     end
 
     def self.next_migration_number(dirname)