+++ spec/dummy_app/app/models/thing.rb
@@ -2,4 +2,6 @@
 
 class Thing < ActiveRecord::Base
   has_paper_trail save_changes: false
+
+  belongs_to :person, optional: true
 end